chore: move test classes to dedicated package name
