+++ core/src/test/java/org/jboss/pnc/rex/test/common/TransitionRecorder.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rex.core.common;
+package org.jboss.pnc.rex.test.common;
 
 import io.vertx.core.impl.ConcurrentHashSet;
 import lombok.extern.slf4j.Slf4j;
