+++ core/src/test/java/org/jboss/pnc/rex/test/endpoints/HttpEndpoint.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.rex.core.endpoints;
+package org.jboss.pnc.rex.test.endpoints;
 
 import io.vertx.mutiny.core.buffer.Buffer;
 import io.vertx.mutiny.ext.web.client.HttpResponse;
